backend/requirements.txt
flask
flask-cors
python-dotenv

langgraph
langchain
chromadb        
sentence-transformers
torch
pypdf
requests

psycopg2-binary
sqlalchemy
python-jose
passlib[bcrypt]
